Execute .ts script files using the ts-node which inherits tsconfig.json unlike tsc
