+++ polaris.shopify.com/tsconfig.json
@@ -32,5 +32,12 @@
     "constants.js",
     "../polaris-tokens/dist"
   ],
-  "exclude": ["node_modules", "scripts/get-props/testData"]
+  "exclude": ["node_modules", "scripts/get-props/testData"],
+  "ts-node": {
+    "transpileOnly": true,
+    "files": true,
+    "compilerOptions": {
+      "module": "CommonJS"
+    }
+  }
 }
